adds tests for search bar display logic * extracts capture_view_locals method to ControllerHelpers
