+++ spec/rails_helper.rb
@@ -20,6 +20,7 @@ end
 RSpec.configure do |config|
   config.include Features, type: :feature
   config.include DashboardHelpers
+  config.include ControllerHelpers
   config.infer_base_class_for_anonymous_controllers = false
   config.infer_spec_type_from_file_location!
   config.use_transactional_fixtures = false
